fix(markdown,spellcheck): platform-split native away from jsdom and harper
Cold Android builds were impossible: sanitize.ts's juice import + jsdom fallback resolve Node builtins under Metro android, and harper.js's inlined WASM was 24.3 MB of a 33.2 MB bundle that Hermes (no WebAssembly/Worker) can never execute — plus an unhandled rejection per lint pass. sanitize.native.ts passes through (RNRH executes nothing and ignores style/script); index.native.ts no-ops the linter behind type-only harper imports. Verified: expo export --platform android now succeeds at 8.7 MB.
